[rename]: 브랜드 카드 컴포넌트 디렉토리 변경 (#139)
기존: components/feature/
변경: layouts/SearchResult/SearchContents/
