+++ src/layouts/SearchResult/SearchContents/BrandCard/index.module.scss
