+++ src/layouts/SearchResult/SearchContents/BrandCard/index.tsx
